Start transfer of image to python icpGui.
This commit only transfers the size of the image.

The plan to use QProcess+QSocketNotifier did not work out well because of gdb
messing with stdin and not being able to open cin in binary mode.

Instead, went with QLocalSocket, which is better anyway.  For transfering the
image pixel buffer, QSharedMemory will probably work nicely.

For now, StreamingImageImporter is based on a float image, but will probably
change it to be templated in the future.