Fix icpmodule cleanup robustness.
After adding ScalarType to the DatagramContent enum,
on exit in _cleanup_QProcess, the icpmodule would do a 30 second timeout on
'waitForFinished()', then another 30 second timeout for similar when calling
delete on the process object.  This would only occur if the 'shell sleep 1'
statement was uncommented.  Why this is so, I am not sure, but the following
change makes everything work properly again.