Fix gdb thinking icpGui is an inferior.
The test that was run previously was printing the image object at the end, then
exiting.  At some points, gdb seemed to be hanging on exit for some reason.
Extending the test has revealed the cause of the problem -- gdb thinks the
icpGui child process is an inferior, gets confused, then gets caught in
sigsuspend().

I was not able to get around this using QProcess::start() or
QProcess::startDetached() with setsid().  Instead, using Python's
subprocess.Popen() seems to work.  According to the Python docs, this launches
the process with '/bin/sh -c'.  Hopefully it works on Windows too.

Everything should be working as it should.  Launching and establishing
inter-process communication with the icpGui requires no interaction from the
user.  There is a unique icpGui instance for every gdb instance.  If the icpGui
is closed during a gdb session, a new one will automatically be generated.